+++ src/client/java/fr/madu59/obe/client/OBEClient.java
@@ -3,6 +3,7 @@ package fr.madu59.obe.client;
 import fr.madu59.obe.OBE;
 import fr.madu59.obe.client.compat.ModCompat;
 import fr.madu59.obe.client.config.configscreen.OptimisedBlockEntitiesConfigScreen;
+import fr.madu59.obe.client.platform.PlatformHelper;
 import fr.madu59.obe.client.registry.Registry;
 import fr.madu59.obe.client.resources.loader.SkullPackLoader;
 import net.fabricmc.api.ClientModInitializer;
@@ -17,6 +18,7 @@ public class OBEClient implements ClientModInitializer {
 		ModCompat.init();
 		OptimisedBlockEntitiesConfigScreen.registerCommand();
 		ResourceLoader.get(PackType.CLIENT_RESOURCES).registerReloadListener(Identifier.tryParse("obe:skulls"), new SkullPackLoader());
+		PlatformHelper.registerPlatformEvents();
 	}
 
 	public static void debug(String debugInfo) {