+++ src/client/java/fr/madu59/obe/client/renderer/blockentity/ext/BlockEntityExt.java
@@ -1,33 +1,8 @@
 package fr.madu59.obe.client.renderer.blockentity.ext;
 
-import fr.madu59.obe.client.renderer.blockentity.misc.RenderModeManager.RenderMode;
+import fr.madu59.obe.client.renderer.entity.ext.EntityExt;
 
-public interface BlockEntityExt {
-    boolean isSupportedBlockEntity();
-    void isSupportedBlockEntity(boolean bl);
-
-    RenderMode renderMode();
-    void renderMode(RenderMode mode);
-
-    RenderMode renderModeDelayed();
-    void renderModeDelayed(RenderMode mode);
-
-    boolean hasSpecialRenderer();
-    void hasSpecialRenderer(boolean bl);
-
-    boolean isTimerFinished();
-    void setTimer(long start, int duration);
-
-    boolean isEnabled();
-    void isEnabled(boolean bl);
-    
+public interface BlockEntityExt extends EntityExt {
     boolean renderBoth();
     void renderBoth(boolean bl);
-
-    boolean shouldSkipBeRendering();
-    void shouldSkipBeRendering(boolean bl);
-    void shouldSkipBeRenderingAndUpdate(boolean bl);
-
-    boolean forceEntity();
-    void forceEntity(boolean bl);
 }