+++ src/client/java/fr/madu59/obe/client/util/blockentity/ShulkerBoxUtil.java
@@ -10,11 +10,10 @@ import net.minecraft.core.Direction;
 import net.minecraft.resources.Identifier;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.level.block.ShulkerBoxBlock;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class ShulkerBoxUtil {
-    public static Identifier getShulkerBoxMaterial(BlockState state){
+public class ShulkerBoxUtil extends AbstractUtil{
+    public static Identifier getMaterial(BlockState state){
         if(state.getBlock() instanceof ShulkerBoxBlock block){
             DyeColor color = block.getColor();
             if (color == null) {
@@ -26,20 +25,12 @@ public class ShulkerBoxUtil {
         else return null;
     }
 
-    public static ModelLayerLocation getShulkerBoxModelLayerLocation(BlockState state){
+    public static ModelLayerLocation getModelLayerLocation(BlockState state){
         return ModelLayers.SHULKER_BOX;
     }
 
-    public static void transformShulkerBox(BlockState state, PoseStack poseStack){
+    public static void transform(BlockState state, PoseStack poseStack){
         Direction facing = state.getValue(ShulkerBoxBlock.FACING);
         poseStack.mulPose(ShulkerBoxRenderer.modelTransform(facing));
-    }
-
-    public static ModelLayerLocation getShulkerBoxModelLayerLocation(BlockState state, BlockEntity be){
-        return getShulkerBoxModelLayerLocation(state);
-    }
-
-    public static void transformShulkerBox(BlockState state, BlockEntity be, PoseStack poseStack){
-        transformShulkerBox(state, poseStack);
     }
 }