+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # What's new:
-- Removed custom skulls from the mod's file, they can be added back with [this](https://modrinth.com/resourcepack/obe-x-hypixel-skyblock) resource pack
-- Added custom skulls textures loading from resource packs
+- Added optimization for cushions
+- Fixed a rendering bug with bells